pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-data_clean_env"
version = "0.1.0"
description = "Data Cleaning Environment for OpenEnv — A real-world data quality simulation for training AI agents"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
server = "data_clean_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["data_clean_env", "data_clean_env.server", "data_clean_env.data"]
package-dir = { "data_clean_env" = ".", "data_clean_env.server" = "server", "data_clean_env.data" = "data" }